HTML, CSS, and JS changes
Added ul to HTML file to house the pokemon index items, added two functions inside the IIFE function. The main function added (addListItem()) displays the 'pokedex' as buttons, and logs the full information of the "clicked" pokemon button name into the console. In CSS, made changes to the design of the button, to make them look cleaner and fit the aesthetic.