Updates to JS
Added polyfills for fetch and promise, for them to be implemented into the main scripts.js file. Also linked them in the index.html file so they do get called. Updated the scripts.js file and added the loadDetails function as well as the showDetails function. Which both play a role in collecting data from the given pokedex api, and to use the api site as the source of data for each of the pokemon being displyed. Functions work so when you click on each displayed pokemon name, their information gathered from the api, get logged in the console.