Updated CSS and JS, added styling and types
For JS, I added the types of each pokemon so they would show up in the modal. I also styled the css so it looks alot more appealing. Made the pictures of each pokemon look better. On the button they are animate, and once the modal is opened, it is their official picture