Finishing touches on app
Customized the styling for the page in css just a bit more, to make it look a bit more realistic. Added linting to make sure there were no bugs. Updated the README file to convey what the page is going to be about and how to navigate it. And minified the JS and CSS.